running refactor, tests need refactoring
